nvidia-gl-G06: Revert providing/obsoleting suse-prime
Early feedback shows that still users rely on suse-prime or on other tools
like supergfxctl, which conflict with suse-prime and therefore conflict now
with nvidia-gl-G06, which makes it impossible to update driver to 570 without
uninstalling the package conflicting with suse-prime.

Nevertheless we won't install suse-prime by default let alone enable and
confiugre it by default in %post.